fixing x64 build issue with num::clamp calls
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@975 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ lass/python/pysequence.h
@@ -152,7 +152,7 @@ namespace impl
 		PyObject* const slice(Py_ssize_t low, Py_ssize_t high, Py_ssize_t step) const
 		{
 			const Py_ssize_t size = this->length();
-			low = num::clamp(low, 0, size);
+			low = num::clamp(low, Py_ssize_t(0), size);
 			high = num::clamp(high, low, size);
 			const Py_ssize_t n = (high - low + step - 1) / step;
 			TPyObjPtr s(PyList_New(n));
@@ -187,7 +187,7 @@ namespace impl
 				return -1;
 			}
 			const Py_ssize_t size = this->length();
-			low = num::clamp(low, 0, size);
+			low = num::clamp(low, Py_ssize_t(0), size);
 			high = num::clamp(high, low, size);
 			TConstContainerPtr b;
 			if (other)
